refactor: 반려동물 관련 로직 pet으로 이동
반려동물 관련된 로직을 pet 패키지에서 관리할 수 있도록 변경
