+++ src/main/java/com/codeit/donggrina/domain/group/controller/GroupController.java
@@ -103,26 +103,4 @@ public class GroupController {
             .message("가족(그룹) 멤버 삭제 성공")
             .build();
     }
-
-    @PostMapping("/my/pets")
-    public ApiResponse<Void> addPet(@AuthenticationPrincipal CustomOAuth2User user,
-        @RequestBody @Validated PetAddRequest petAddRequest) {
-
-        groupService.addPet(user.getMemberId(), petAddRequest);
-        return ApiResponse.<Void>builder()
-            .code(HttpStatus.OK.value())
-            .message("반려동물 등록 성공")
-            .build();
-    }
-
-    @GetMapping("/my/pets")
-    public ApiResponse<List<PetFindListResponse>> findPetList(
-        @AuthenticationPrincipal CustomOAuth2User user) {
-
-        return ApiResponse.<List<PetFindListResponse>>builder()
-            .code(HttpStatus.OK.value())
-            .message("반려동물 전체 조회 성공")
-            .data(groupService.findPetList(user.getMemberId()))
-            .build();
-    }
 }
